requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-core>=0.3.0
groq>=0.11.0
yfinance>=0.2.40
python-dotenv>=1.0.0
streamlit>=1.38.0
plotly>=5.22.0
pandas>=2.2.0
finnhub-python>=2.4.19
